Update mapDeserializer to match model classes

+++ src/model/Intersection.java
@@ -2,23 +2,23 @@ package model;
 
 public class Intersection {
 
-	private float latitude;
-	private float longitude;
+	private double latitude;
+	private double longitude;
 	
 	private long id;
 
-	public Intersection(float latitude, float longitude, int id) {
+	public Intersection(double latitude, double longitude, long id) {
 		super();
 		this.latitude = latitude;
 		this.longitude = longitude;
 		this.id = id;
 	}
 
-	public float getLatitude() {
+	public double getLatitude() {
 		return latitude;
 	}
 
-	public float getLongitude() {
+	public double getLongitude() {
 		return longitude;
 	}
